Set size, stride, and offset of functional tensor.
Fix: #120585

This PR modifies at::functionalization::impl::set_sizes_strides_offset, calling
TensorImpl::set_sizes_and_strides not only for the FunctionalTensorWrapper tensor, but
also for its inner value (which is also a tensor).

The rationale is that FunctionalTensorWrapper actually overrides its metadata getter
methods, forwarding them to its inner value. Therefore, whenever we try to set that
information for a FunctionalTensorWrapper, it makes sense to actually set it for its
inner value.

ghstack-source-id: ef8a599ebdd3ee7116172b9fdb25b1346b9a7ad6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/135237